Move Spinlock to sync module
